+++ flutter-hms-scan/android/src/main/java/com/huawei/hms/flutter/scan/scanutils/ScanUtilsMethodCallHandler.java
@@ -21,12 +21,16 @@ import static android.app.Activity.RESULT_OK;
 import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Bitmap;
+import android.os.Parcelable;
 import android.text.TextUtils;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
 
+import com.huawei.hms.flutter.scan.ScanPlugin;
 import com.huawei.hms.flutter.scan.logger.HMSLogger;
+import com.huawei.hms.flutter.scan.multiprocessor.MultiProcessorActivity;
+import com.huawei.hms.flutter.scan.utils.Constants;
 import com.huawei.hms.flutter.scan.utils.Errors;
 import com.huawei.hms.flutter.scan.utils.ValueGetter;
 import com.huawei.hms.hmsscankit.ScanUtil;
@@ -34,6 +38,9 @@ import com.huawei.hms.hmsscankit.WriterException;
 import com.huawei.hms.ml.scan.HmsBuildBitmapOption;
 import com.huawei.hms.ml.scan.HmsScan;
 import com.huawei.hms.ml.scan.HmsScanAnalyzerOptions;
+import com.huawei.hms.ml.scan.HmsScanFrame;
+import com.huawei.hms.ml.scan.HmsScanFrameOptions;
+import com.huawei.hms.ml.scan.HmsScanResult;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -46,8 +53,18 @@ import java.io.ByteArrayOutputStream;
 import java.util.List;
 
 public class ScanUtilsMethodCallHandler
-    implements MethodChannel.MethodCallHandler, PluginRegistry.ActivityResultListener {
+        implements MethodChannel.MethodCallHandler, PluginRegistry.ActivityResultListener {
     private static final int REQUEST_CODE_SCAN_DEFAULT = 13;
+
+    private static final int REQUEST_CODE_SCAN_BITMAP = 0X02;
+
+    private static final int REQUEST_CODE_SCAN_DECODE = 0X03;
+
+    public static final int SCANMODEDECODE = 222;
+
+    public static final int SCANMODEDECODEWITHBITMAP = 333;
+
+    private int channelId;
 
     private final HMSLogger mHMSLogger;
 
@@ -57,7 +74,9 @@ public class ScanUtilsMethodCallHandler
 
     private Gson gson;
 
-    public ScanUtilsMethodCallHandler(final Activity activity) {
+    public ScanUtilsMethodCallHandler(final Activity activity, MethodChannel channel) {
+        channelId = hashCode();
+        ScanPlugin.SCAN_CHANNELS.put(channelId, channel);
         mActivity = activity;
         gson = new GsonBuilder().setPrettyPrinting().create();
         mHMSLogger = HMSLogger.getInstance(mActivity.getApplicationContext());
@@ -75,6 +94,9 @@ public class ScanUtilsMethodCallHandler
                 break;
             case "decodeWithBitmap":
                 decodeWithBitmap(call, result);
+                break;
+            case "decode":
+                decode(call, result);
                 break;
             case "disableLogger":
                 mHMSLogger.disableLogger();
@@ -119,9 +141,9 @@ public class ScanUtilsMethodCallHandler
         } else {
             Log.i("DefaultView", Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorMessage());
             result.error(Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorCode(),
-                Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorMessage(), null);
+                    Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorMessage(), null);
             mHMSLogger.sendSingleEvent("ScanUtilsMethodCallHandler.defaultView",
-                Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorCode());
+                    Errors.SCAN_UTIL_NO_CAMERA_PERMISSION.getErrorCode());
         }
     }
 
@@ -174,9 +196,14 @@ public class ScanUtilsMethodCallHandler
 
     private void decodeWithBitmap(MethodCall call, MethodChannel.Result result) {
         // Arguments from call
+        pendingResult = result;
         int scanType = ValueGetter.getInt("scanType", call);
         List<Integer> additionalScanTypes = call.argument("additionalScanTypes");
         Bitmap bitmap = ValueGetter.bitmapForDecoders(call, "data");
+        boolean photoMode = ValueGetter.getBoolean("photoMode", call);
+        HmsScanAnalyzerOptions.Creator creator = new HmsScanAnalyzerOptions.Creator();
+        HmsScanAnalyzerOptions options = creator.create();
+        HmsScan[] hmsScans = ScanUtil.decodeWithBitmap(mActivity.getApplicationContext(), bitmap, options);
 
         // List<Integer> to int[]
         int[] scanTypesIntArray = null;
@@ -184,26 +211,106 @@ public class ScanUtilsMethodCallHandler
             scanTypesIntArray = ValueGetter.scanTypesListToArray(additionalScanTypes);
         }
 
-        // Default view options
-        HmsScanAnalyzerOptions.Creator creator = new HmsScanAnalyzerOptions.Creator();
-        creator.setHmsScanTypes(scanType, scanTypesIntArray);
-        creator.setPhotoMode(true);
-        HmsScanAnalyzerOptions options = creator.create();
+        if (!photoMode) {
+            Intent intent = new Intent(mActivity.getApplicationContext(), MultiProcessorActivity.class);
 
-        // Decode with bitmap
-        mHMSLogger.startMethodExecutionTimer("ScanUtilsMethodCallHandler.decodeWithBitmap");
-        HmsScan[] hmsScans = ScanUtil.decodeWithBitmap(mActivity, bitmap, options);
-        mHMSLogger.sendSingleEvent("ScanUtilsMethodCallHandler.decodeWithBitmap");
+            intent.putExtra(Constants.CHANNEL_ID_KEY, channelId);
+            intent.putExtra("scanType", scanType);
+            if (additionalScanTypes != null) {
+                intent.putExtra("additionalScanTypes", scanTypesIntArray);
+            }
 
-        // Send result to flutter
-        if (hmsScans != null && hmsScans.length > 0 && hmsScans[0] != null && !TextUtils.isEmpty(
-            hmsScans[0].getOriginalValue())) {
-            result.success(gson.toJson(hmsScans[0]));
+            intent.putExtra("scanMode", SCANMODEDECODEWITHBITMAP);
+            mActivity.startActivityForResult(intent, REQUEST_CODE_SCAN_BITMAP);
         } else {
-            result.error(Errors.DECODE_WITH_BITMAP_ERROR.getErrorCode(), Errors.DECODE_WITH_BITMAP_ERROR.getErrorMessage(),
-                null);
+            if (bitmap.toString().isEmpty()) {
+                pendingResult.error(Errors.MP_CHANNEL_ERROR.getErrorCode(), Errors.MP_CHANNEL_ERROR.getErrorMessage(),
+                        null);
+            }
+            // Default view options
+            creator.setHmsScanTypes(scanType, scanTypesIntArray);
+            creator.setPhotoMode(true);
+
+            // Decode with bitmap
+            mHMSLogger.startMethodExecutionTimer("ScanUtilsMethodCallHandler.decodeWithBitmap");
+            mHMSLogger.sendSingleEvent("ScanUtilsMethodCallHandler.decodeWithBitmap");
+
+            // Send result to flutter
+            if (hmsScans != null && hmsScans.length > 0 && hmsScans[0] != null
+                    && !TextUtils.isEmpty(hmsScans[0].getOriginalValue())) {
+                result.success(gson.toJson(hmsScans));
+            } else {
+                result.error(Errors.DECODE_WITH_BITMAP_ERROR.getErrorCode(),
+                        Errors.DECODE_WITH_BITMAP_ERROR.getErrorMessage(),
+                        null);
+            }
+
+            if (bitmap.toString().isEmpty()) {
+                pendingResult.error(Errors.MP_CHANNEL_ERROR.getErrorCode(), Errors.MP_CHANNEL_ERROR.getErrorMessage(),
+                        null);
+            }
         }
 
+    }
+
+    private void decode(MethodCall call, MethodChannel.Result result) {
+
+        pendingResult = result;
+        int scanType = ValueGetter.getInt("scanType", call);
+        List<Integer> additionalScanTypes = call.argument("additionalScanTypes");
+        boolean photoMode = ValueGetter.getBoolean("photoMode", call);
+        boolean multiMode = ValueGetter.getBoolean("multiMode", call);
+        boolean parseResult = ValueGetter.getBoolean("parseResult", call);
+
+        // List<Integer> to int[]
+        int[] scanTypesIntArray = null;
+        if (additionalScanTypes != null) {
+            scanTypesIntArray = ValueGetter.scanTypesListToArray(additionalScanTypes);
+        }
+
+        if (!photoMode) {
+            Intent intent = new Intent(mActivity.getApplicationContext(), MultiProcessorActivity.class);
+
+            intent.putExtra("scanType", scanType);
+            intent.putExtra("multiMode", multiMode);
+            intent.putExtra("parseResult", parseResult);
+            if (additionalScanTypes != null) {
+                intent.putExtra("additionalScanTypes", scanTypesIntArray);
+            }
+            intent.putExtra("scanMode", SCANMODEDECODE);
+            intent.putExtra(Constants.CHANNEL_ID_KEY, channelId);
+            mActivity.startActivityForResult(intent, REQUEST_CODE_SCAN_DECODE);
+        } else {
+            Bitmap bitmap = ValueGetter.bitmapForDecoders(call, "data");
+            if (bitmap.toString().isEmpty()) {
+                pendingResult.error(Errors.MP_CHANNEL_ERROR.getErrorCode(), Errors.MP_CHANNEL_ERROR.getErrorMessage(),
+                        null);
+            }
+
+            HmsScanFrame frame;
+            frame = new HmsScanFrame(bitmap);
+
+            HmsScanFrameOptions.Creator creator = new HmsScanFrameOptions.Creator();
+            creator.setHmsScanTypes(scanType, scanTypesIntArray);
+            creator.setPhotoMode(true);
+            creator.setMultiMode(multiMode);
+            creator.setParseResult(parseResult);
+            HmsScanFrameOptions options = creator.create();
+
+            mHMSLogger.startMethodExecutionTimer("ScanUtilsMethodCallHandler.decode");
+            HmsScanResult scanResult = ScanUtil.decode(mActivity.getApplicationContext(), frame, options);
+            HmsScan[] hmsScans = scanResult.getHmsScans();
+            mHMSLogger.sendSingleEvent("ScanUtilsMethodCallHandler.decode");
+
+            if (hmsScans != null && hmsScans.length > 0 && hmsScans[0] != null
+                    && !TextUtils.isEmpty(hmsScans[0].getOriginalValue())) {
+                result.success(gson.toJson(hmsScans));
+
+            } else {
+                result.error(Errors.DECODE_WITH_BITMAP_ERROR.getErrorCode(),
+                        Errors.DECODE_WITH_BITMAP_ERROR.getErrorMessage(), null);
+            }
+        }
     }
 
     @Override
@@ -215,7 +322,7 @@ public class ScanUtilsMethodCallHandler
         // Default View
         if (requestCode == REQUEST_CODE_SCAN_DEFAULT) {
             int errorCode = data.getIntExtra(ScanUtil.RESULT_CODE, ScanUtil.SUCCESS);
-            if (errorCode == ScanUtil.SUCCESS){
+            if (errorCode == ScanUtil.SUCCESS) {
                 if (pendingResult != null) {
                     HmsScan obj = data.getParcelableExtra(ScanUtil.RESULT);
                     // Sending Result
@@ -224,6 +331,26 @@ public class ScanUtilsMethodCallHandler
                             .sendSingleEvent("ScanUtilsMethodCallHandler.defaultView");
                     pendingResult = null; // reset
                 }
+            } else if (errorCode == ScanUtil.SCAN_NO_DETECTED) {
+                HMSLogger.getInstance(mActivity.getApplicationContext()).sendSingleEvent("ScanUtilsMethodCallHandler",
+                        "null data");
+                pendingResult.error(Errors.SCAN_NO_DETECTED.getErrorCode(), "No barcode is detected", null);
+            }
+        } else if (requestCode == REQUEST_CODE_SCAN_DECODE || requestCode == REQUEST_CODE_SCAN_BITMAP) {
+            int errorCode = data.getIntExtra(ScanUtil.RESULT_CODE, ScanUtil.SUCCESS);
+            if (errorCode == ScanUtil.SUCCESS) {
+                if (pendingResult != null) {
+                    Parcelable[] obj = data.getParcelableArrayExtra(ScanUtil.RESULT);
+                    // Sending Result
+                    pendingResult.success(gson.toJson(obj));
+                    HMSLogger.getInstance(mActivity.getApplicationContext())
+                            .sendSingleEvent("ScanUtilsMethodCallHandler.decodeWithBitmap");
+                    pendingResult = null; // reset
+                }
+            } else if (errorCode == ScanUtil.SCAN_NO_DETECTED) {
+                HMSLogger.getInstance(mActivity.getApplicationContext()).sendSingleEvent("ScanUtilsMethodCallHandler",
+                        "null data");
+                pendingResult.error(Errors.SCAN_NO_DETECTED.getErrorCode(), "No barcode is detected", null);
             }
         } else {
             pendingResult = null;